huge electron cleanup
