+++ electron/src/appListeners.ts
@@ -1,8 +1,25 @@
 import { app, BrowserWindow } from "electron";
-import { createMainWindow, settings } from "./main";
-import { windows } from "./helpers/globalState";
+import { isLinux, settings, windows } from "./helpers/globalState";
+import { createUpdaterSplashWindow, skipUpdateCheck } from "./updaterListeners";
+import isDev from 'electron-is-dev'
+import { autoUpdater } from 'electron-updater'
+import { createMainWindow } from "./helpers/utils";
 
 export const startAppListeners = () => {
+
+    // app entry point
+    // creates splash window to look for updates and then start the main window
+    app.on('ready', () => {
+        createUpdaterSplashWindow()
+        settings.loadSettingsFromDb().then(async (settings) => {
+            autoUpdater.autoDownload = settings.shouldAutoUpdate
+            autoUpdater.allowPrerelease = settings.allowPreRelease
+            if (!windows.splash) return
+            if (isDev || isLinux || !settings.shouldAutoUpdate) skipUpdateCheck(windows.splash)
+            if (!isDev && !isLinux) await autoUpdater.checkForUpdates()
+        })
+    })
+
     app.on("second-instance", async () => {
         if (windows.mainWindow) {
             if (windows.mainWindow.isDestroyed()) {
@@ -14,7 +31,7 @@ export const startAppListeners = () => {
         } else {
             await createMainWindow();
         }
-    });
+    })
     
     app.on('window-all-closed', () => {
         if (!settings.shouldMinimizeToTray) app.quit()
@@ -22,6 +39,6 @@ export const startAppListeners = () => {
     
     app.on("activate", function () {
         if (BrowserWindow.getAllWindows().length === 0) createMainWindow();
-    });
+    })
     
 }
