+++ electron/src/helpers/controllers/c-index.controller.ts
@@ -1,11 +1,10 @@
 import { ipcMain } from 'electron';
 import { uniqueId } from 'lodash';
-import { createMainWindow, kkStateController, windows } from '../../main';
-import { shared, userType } from '../../shared';
 
 import { Body, Controller, Get, Post, Header, Route, Tags, Response, Security } from 'tsoa';
-import { GenericResponse, PairBody, PairResponse } from '../types';
-import { db } from '../globalState';
+import { GenericResponse, PairBody, PairResponse, UserType } from '../types';
+import { db, kkStateController, shared, windows } from '../globalState';
+import { createMainWindow } from '../utils';
 
 
 export class ApiError extends Error {
@@ -132,9 +131,7 @@ export class CIndexController extends Controller {
     @Get('/user')
     @Security("api_key")
     @Response(401, "Please provide a valid serviceKey")
-    public async user(): Promise<userType> {
+    public async user(): Promise<UserType> {
         return shared.USER
     }
-
-
 }
