+++ electron/src/helpers/controllers/d-pubkey-controller.ts
@@ -1,13 +1,12 @@
 import { ipcMain } from 'electron';
-import { kkStateController, windows } from '../../main';
 import { Body, Controller, Post, Security, Route, Tags, Response } from 'tsoa';
 import { GetPublicKey } from '../types';
 import { EosGetPublicKey, RippleGetAddress, BinanceGetAddress, ETHGetAddress, OsmosisGetAddress, PublicKey, ThorchainGetAddress, CosmosGetAddress, BTCGetAddress } from '@shapeshiftoss/hdwallet-core'
 import { checkKeepKeyUnlocked } from '../utils';
+import { kkStateController, windows } from '../globalState';
 @Tags('KeepKey Wallet Endpoints')
 @Route('')
 export class DPubkeyController extends Controller {
-
     @Post('/getPublicKeys')
     @Security("api_key")
     @Response(500, "Internal server error")
