+++ electron/src/helpers/controllers/f-developer-controller.ts
@@ -2,7 +2,7 @@ import { Body, Controller, Post, Security, Route, Tags, Response } from 'tsoa';
 import wait from 'wait-promise'
 import { ResetDevice, LoadDevice, ETHSignedTx } from '@shapeshiftoss/hdwallet-core'
 import { checkKeepKeyUnlocked } from '../utils';
-import { kkStateController } from '../../main';
+import { kkStateController } from '../globalState';
 
 export type policy = {
     policyName?: string,
