+++ electron/src/helpers/controllers/g-recovery-controller.ts
@@ -1,8 +1,8 @@
 import { Body, Controller, Post, Security, Route, Tags, Response } from 'tsoa';
 import wait from 'wait-promise'
 import { RecoverDevice, ETHSignedTx } from '@shapeshiftoss/hdwallet-core'
-import { checkKeepKeyUnlocked } from '../utils';
-import { kkStateController } from '../../main';
+import { checkKeepKeyUnlocked } from '../utils'
+import { kkStateController } from '../globalState'
 
 @Tags('Recovery Endpoints')
 @Route('')
