+++ electron/src/helpers/controllers/h-device-info-controller.ts
@@ -2,7 +2,7 @@ import { Controller, Get, Security, Route, Tags, Response } from 'tsoa';
 import wait from 'wait-promise'
 import { ETHSignedTx } from '@shapeshiftoss/hdwallet-core'
 import { checkKeepKeyUnlocked } from '../utils'
-import { kkStateController } from '../../main'
+import { kkStateController } from '../globalState';
 
 @Tags('Device Info Endpoints')
 @Route('')
