+++ electron/src/updaterListeners.ts
@@ -1,30 +1,17 @@
 import { autoUpdater } from 'electron-updater'
-import { createMainWindow, settings, shouldShowWindow } from '../main';
 import isDev from 'electron-is-dev'
-import { app, BrowserWindow, ipcMain } from 'electron';
-import log from 'electron-log';
-import path from 'path';
-import { isLinux, windows } from './globalState';
+import { app, BrowserWindow, ipcMain } from 'electron'
+import log from 'electron-log'
+import path from 'path'
+import { isLinux, settings, shouldShowWindow, windows } from './helpers/globalState'
+import { createMainWindow } from './helpers/utils'
 
 let skipUpdateTimeout: NodeJS.Timeout;
 let windowShowInterval: NodeJS.Timeout
 export let skipUpdateCheckCompleted = false
 
-export const setupAutoUpdater = () => {
+export const startUpdaterListeners = () => {
     autoUpdater.logger = log;
-
-    app.on('ready', () => {
-        createSplashWindow()
-
-        settings.loadSettingsFromDb().then(async (settings) => {
-            autoUpdater.autoDownload = settings.shouldAutoUpdate
-            autoUpdater.allowPrerelease = settings.allowPreRelease
-
-            if (!windows.splash) return
-            if (isDev || isLinux || !settings.shouldAutoUpdate) skipUpdateCheck(windows.splash)
-            if (!isDev && !isLinux) await autoUpdater.checkForUpdates()
-        })
-    })
 
     autoUpdater.on("update-available", (info) => {
         if (skipUpdateCheckCompleted) return
@@ -34,49 +21,44 @@ export const setupAutoUpdater = () => {
         skipUpdateTimeout = setTimeout(() => {
             if (!windows.splash) return
             skipUpdateCheck(windows.splash);
-        }, 60000);
-    });
-
+        }, 60000)
+    })
 
     autoUpdater.on("download-progress", (progress) => {
-        let prog = Math.floor(progress.percent);
-        if (windows.splash && !windows.splash.isDestroyed()) windows.splash.webContents.send("@update/percentage", prog);
+        let prog = Math.floor(progress.percent)
+        if (windows.splash && !windows.splash.isDestroyed()) windows.splash.webContents.send("@update/percentage", prog)
         if (windows.splash && !windows.splash.isDestroyed()) windows.splash.setProgressBar(prog / 100);
-        if (windows.mainWindow && !windows.mainWindow.isDestroyed()) windows.mainWindow.webContents.send("@update/percentage", prog);
-        if (windows.mainWindow && !windows.mainWindow.isDestroyed()) windows.mainWindow.setProgressBar(prog / 100);
+        if (windows.mainWindow && !windows.mainWindow.isDestroyed()) windows.mainWindow.webContents.send("@update/percentage", prog)
+        if (windows.mainWindow && !windows.mainWindow.isDestroyed()) windows.mainWindow.setProgressBar(prog / 100)
         // stop timeout that skips the update
         if (skipUpdateTimeout) {
-            clearTimeout(skipUpdateTimeout);
+            clearTimeout(skipUpdateTimeout)
         }
-    });
-
+    })
 
     autoUpdater.on("update-downloaded", () => {
         if (skipUpdateCheckCompleted) return
         if (windows.splash) windows.splash.webContents.send("@update/relaunch");
         // stop timeout that skips the update
         if (skipUpdateTimeout) {
-            clearTimeout(skipUpdateTimeout);
+            clearTimeout(skipUpdateTimeout)
         }
         setTimeout(() => {
-            autoUpdater.quitAndInstall();
-        }, 1000);
-    });
-
+            autoUpdater.quitAndInstall()
+        }, 1000)
+    })
 
     autoUpdater.on("update-not-available", () => {
         if (skipUpdateCheckCompleted) return
         if (!windows.splash) return
-        skipUpdateCheck(windows.splash);
-    });
-
+        skipUpdateCheck(windows.splash)
+    })
 
     autoUpdater.on("error", () => {
         if (skipUpdateCheckCompleted) return
         if (!windows.splash) return
-        skipUpdateCheck(windows.splash);
-    });
-
+        skipUpdateCheck(windows.splash)
+    })
 
     ipcMain.on('@app/update', async (event) => {
         if (isDev) return event.sender.send('@app/update', { updateInfo: { version: app.getVersion() } })
@@ -95,12 +77,11 @@ export const setupAutoUpdater = () => {
     })
 }
 
-
 export const skipUpdateCheck = (splash: BrowserWindow) => {
     createMainWindow();
-    splash.webContents.send("@update/notfound");
+    splash.webContents.send("@update/notfound")
     if (isLinux || isDev) {
-        splash.webContents.send("@update/skipCheck");
+        splash.webContents.send("@update/skipCheck")
     }
     // stop timeout that skips the update
     if (skipUpdateTimeout) {
@@ -129,7 +110,7 @@ export const skipUpdateCheck = (splash: BrowserWindow) => {
     skipUpdateCheckCompleted = true
 }
 
-export const createSplashWindow = () => {
+export const createUpdaterSplashWindow = () => {
     windows.splash = new BrowserWindow({
         width: 300,
         height: 410,
